webxr/ar-module/xrSession_interactionMode.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="../resources/webxr_util.js"></script>
  <script src="../resources/webxr_test_constants.js"></script>
  <canvas></canvas>
  <script>
    xr_session_promise_test(
      "Tests interactionMode for an AR device",
      (session) => {
        assert_equals(session.interactionMode, "world-space");
      }, IMMERSIVE_AR_DEVICE, 'immersive-ar', {});

    xr_session_promise_test(
      "Tests interactionMode for a VR device",
      (session) => {
        assert_equals(session.interactionMode, "world-space");
      }, TRACKED_IMMERSIVE_DEVICE, 'immersive-vr', {});

    xr_session_promise_test(
      "Tests interactionMode for a non-immersive device",
      (session) => {
        assert_equals(session.interactionMode, "screen-space");
      }, VALID_NON_IMMERSIVE_DEVICE, 'inline', {});
</script>
</body>
